+++ test_unstructured_ingest/test-ingest-github.sh
@@ -25,10 +25,11 @@ fi
 
 #shellcheck disable=SC2086
 PYTHONPATH=. ./unstructured/ingest/main.py \
-    --metadata-exclude filename \
+    --metadata-exclude filename,file_directory \
     --github-url dcneiner/Downloadify \
     --git-file-glob '*.html,*.txt' \
     --structured-output-dir github-downloadify-output \
+    --partition-strategy hi_res \
     --reprocess \
     --preserve-downloads \
     --verbose $ACCESS_TOKEN_FLAGS