+++ test_unstructured_ingest/test-ingest-s3.sh
@@ -12,11 +12,12 @@ if [[ "$(find test_unstructured_ingest/expected-structured-output/s3-small-batch
 fi
 
 PYTHONPATH=. ./unstructured/ingest/main.py \
-    --metadata-exclude filename \
+    --metadata-exclude filename,file_directory \
     --s3-url s3://utic-dev-tech-fixtures/small-pdf-set/ \
     --s3-anonymous \
     --structured-output-dir s3-small-batch-output \
     --preserve-downloads \
+    --partition-strategy hi_res \
     --reprocess
 
 OVERWRITE_FIXTURES=${OVERWRITE_FIXTURES:-false}