+++ unstructured/documents/elements.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import datetime
 import hashlib
 import os
@@ -16,7 +18,9 @@ class NoID(ABC):
 @dataclass
 class ElementMetadata:
     filename: Optional[str] = None
+    file_directory: Optional[str] = None
     date: Optional[str] = None
+    filetype: Optional[str] = None
 
     # Page numbers currenlty supported for PDF, HTML and PPT documents
     page_number: Optional[int] = None
@@ -37,7 +41,9 @@ class ElementMetadata:
             self.filename = str(self.filename)
 
         if self.filename is not None:
-            self.filename = os.path.basename(self.filename)
+            file_directory, filename = os.path.split(self.filename)
+            self.file_directory = file_directory or None
+            self.filename = filename
 
     def to_dict(self):
         return {key: value for key, value in self.__dict__.items() if value is not None}
@@ -45,6 +51,12 @@ class ElementMetadata:
     @classmethod
     def from_dict(cls, input_dict):
         return cls(**input_dict)
+
+    def merge(self, other: ElementMetadata):
+        for k in self.__dict__:
+            if getattr(self, k) is None:
+                setattr(self, k, getattr(other, k))
+        return self
 
     def get_date(self) -> Optional[datetime.datetime]:
         """Converts the date field to a datetime object."""