+++ unstructured/file_utils/filetype.py
@@ -1,11 +1,14 @@
+import inspect
 import os
 import re
 import zipfile
 from enum import Enum
-from typing import IO, Optional
+from functools import wraps
+from typing import IO, Callable, List, Optional
 
+from unstructured.documents.elements import Element, PageBreak
 from unstructured.nlp.patterns import LIST_OF_DICTS_PATTERN
-from unstructured.partition.common import exactly_one
+from unstructured.partition.common import _add_element_metadata, exactly_one
 
 try:
     import magic
@@ -17,52 +20,9 @@ except ImportError:  # pragma: nocover
 from unstructured.logger import logger
 from unstructured.nlp.patterns import EMAIL_HEAD_RE
 
-DOCX_MIME_TYPES = [
-    "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
-]
-
-DOC_MIME_TYPES = [
-    "application/msword",
-]
-
-ODT_MIME_TYPES = [
-    "application/vnd.oasis.opendocument.text",
-]
-
-XLSX_MIME_TYPES = [
-    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
-]
-
-XLS_MIME_TYPES = [
-    "application/vnd.ms-excel",
-]
-
-PPTX_MIME_TYPES = [
-    "application/vnd.openxmlformats-officedocument.presentationml.presentation",
-]
-
-PPT_MIME_TYPES = [
-    "application/vnd.ms-powerpoint",
-]
-
-MSG_MIME_TYPES = [
-    "application/vnd.ms-outlook",
-    "application/x-ole-storage",
-]
-
 TXT_MIME_TYPES = [
     "text/plain",
     "message/rfc822",  # ref: https://www.rfc-editor.org/rfc/rfc822
-]
-
-MD_MIME_TYPES = [
-    "text/markdown",
-    "text/x-markdown",
-]
-
-EPUB_MIME_TYPES = [
-    "application/epub",
-    "application/epub+zip",
 ]
 
 # NOTE(robinson) - .docx.xlsx files are actually zip file with a .docx/.xslx extension.
@@ -74,7 +34,6 @@ EXPECTED_DOCX_FILES = [
 ]
 
 EXPECTED_XLSX_FILES = [
-    "docProps/core.xml",
     "xl/workbook.xml",
 ]
 
@@ -132,10 +91,14 @@ STR_TO_FILETYPE = {
     "application/vnd.openxmlformats-officedocument.wordprocessingml.document": FileType.DOCX,
     "image/jpeg": FileType.JPG,
     "image/png": FileType.PNG,
+    "text/plain": FileType.TXT,
     "text/markdown": FileType.MD,
     "text/x-markdown": FileType.MD,
     "application/epub": FileType.EPUB,
     "application/epub+zip": FileType.EPUB,
+    "application/json": FileType.JSON,
+    "application/rtf": FileType.RTF,
+    "text/rtf": FileType.RTF,
     "text/html": FileType.HTML,
     "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": FileType.XLSX,
     "application/vnd.ms-excel": FileType.XLS,
@@ -143,8 +106,14 @@ STR_TO_FILETYPE = {
     "application/vnd.ms-powerpoint": FileType.PPT,
     "application/xml": FileType.XML,
     "application/vnd.oasis.opendocument.text": FileType.ODT,
+    "message/rfc822": FileType.EML,
+    "application/x-ole-storage": FileType.MSG,
+    "application/vnd.ms-outlook": FileType.MSG,
 }
 
+FILETYPE_TO_MIMETYPE = {
+    v: k for k, v in STR_TO_FILETYPE.items() if k not in ("text/x-markdown", "application/epub+zip")
+}
 
 EXT_TO_FILETYPE = {
     ".pdf": FileType.PDF,
@@ -155,6 +124,7 @@ EXT_TO_FILETYPE = {
     ".text": FileType.TXT,
     ".eml": FileType.EML,
     ".xml": FileType.XML,
+    ".htm": FileType.HTML,
     ".html": FileType.HTML,
     ".md": FileType.MD,
     ".xlsx": FileType.XLSX,
@@ -195,13 +165,9 @@ def detect_filetype(
         extension = extension.lower()
         if os.path.isfile(_filename) and LIBMAGIC_AVAILABLE:
             mime_type = magic.from_file(filename or file_filename, mime=True)  # type: ignore
-            # NOTE(crag): for older versions of the OS libmagic package, such as is currently
-            # installed on the Unstructured docker image, .json files resolve to "text/plain"
-            # rather than "application/json". this corrects for that case.
-            if mime_type == "text/plain" and extension == ".json":
-                return FileType.JSON
         else:
             return EXT_TO_FILETYPE.get(extension.lower(), FileType.UNK)
+
     elif file is not None:
         extension = None
         # NOTE(robinson) - the python-magic docs recommend reading at least the first 2048 bytes
@@ -218,50 +184,19 @@ def detect_filetype(
     else:
         raise ValueError("No filename, file, nor file_filename were specified.")
 
-    if mime_type == "application/pdf":
-        return FileType.PDF
+    """Mime type special cases."""
 
-    elif mime_type == "application/json":
+    # NOTE(crag): for older versions of the OS libmagic package, such as is currently
+    # installed on the Unstructured docker image, .json files resolve to "text/plain"
+    # rather than "application/json". this corrects for that case.
+    if mime_type == "text/plain" and extension == ".json":
         return FileType.JSON
 
-    elif mime_type in DOCX_MIME_TYPES:
-        return FileType.DOCX
-
-    elif mime_type in DOC_MIME_TYPES:
-        return FileType.DOC
-
-    elif mime_type in ODT_MIME_TYPES:
-        return FileType.ODT
-
-    elif mime_type in MSG_MIME_TYPES:
-        return FileType.MSG
-
-    elif mime_type == "image/jpeg":
-        return FileType.JPG
-
-    elif mime_type == "image/png":
-        return FileType.PNG
-
-    elif mime_type in MD_MIME_TYPES:
-        # NOTE - I am not sure whether libmagic ever returns these mimetypes.
-        return FileType.MD
-
-    elif mime_type in EPUB_MIME_TYPES:
-        return FileType.EPUB
-
-    # NOTE(robinson) - examples are application/rtf or text/rtf.
-    # magic often returns text/plain for RTF files
-    elif mime_type.endswith("rtf"):
-        return FileType.RTF
-
     elif mime_type.endswith("xml"):
-        if extension and extension == ".html":
+        if extension and (extension == ".html" or extension == ".htm"):
             return FileType.HTML
         else:
             return FileType.XML
-
-    elif mime_type == "text/html":
-        return FileType.HTML
 
     elif mime_type in TXT_MIME_TYPES or mime_type.startswith("text"):
         if extension and extension == ".eml":
@@ -270,25 +205,20 @@ def detect_filetype(
             return FileType.MD
         elif extension and extension == ".rtf":
             return FileType.RTF
+        elif extension and extension == ".html":
+            return FileType.HTML
 
         if _is_text_file_a_json(file=file, filename=filename):
             return FileType.JSON
 
         if file and not extension and _check_eml_from_buffer(file=file) is True:
             return FileType.EML
+
+        # Safety catch
+        if mime_type in STR_TO_FILETYPE:
+            return STR_TO_FILETYPE[mime_type]
+
         return FileType.TXT
-
-    elif mime_type in XLSX_MIME_TYPES:
-        return FileType.XLSX
-
-    elif mime_type in XLS_MIME_TYPES:
-        return FileType.XLS
-
-    elif mime_type in PPTX_MIME_TYPES:
-        return FileType.PPTX
-
-    elif mime_type in PPT_MIME_TYPES:
-        return FileType.PPT
 
     elif mime_type == "application/octet-stream":
         if file and not extension:
@@ -309,6 +239,10 @@ def detect_filetype(
             return EXT_TO_FILETYPE.get(extension.lower(), FileType.ZIP)
         else:
             return EXT_TO_FILETYPE.get(extension.lower(), filetype)
+
+    # For everything else
+    elif mime_type in STR_TO_FILETYPE:
+        return STR_TO_FILETYPE[mime_type]
 
     logger.warning(
         f"The MIME type{f' of {filename!r}' if filename else ''} is {mime_type!r}. "
@@ -372,3 +306,61 @@ def _check_eml_from_buffer(file: IO) -> bool:
         file_head = file_content
 
     return EMAIL_HEAD_RE.match(file_head) is not None
+
+
+def document_to_element_list(
+    document,
+    include_page_breaks: bool = False,
+) -> List[Element]:
+    """Converts a DocumentLayout object to a list of unstructured elements."""
+    elements: List[Element] = []
+    image_formats: List[str] = []
+    num_pages = len(document.pages)
+    for i, page in enumerate(document.pages):
+        for element in page.elements:
+            elements.append(element)
+            if hasattr(page, "image"):
+                image_formats.append(page.image.format)
+        if include_page_breaks and i < num_pages - 1:
+            elements.append(PageBreak())
+
+    if image_formats and all(image_format == "PNG" for image_format in image_formats):
+        filetype = FileType.PNG.name
+    elif image_formats and all(image_format == "JPEG" for image_format in image_formats):
+        filetype = FileType.JPG.name
+    else:
+        filetype = None
+    elements = _add_element_metadata(
+        elements,
+        include_page_breaks=include_page_breaks,
+        filetype=filetype,
+    )
+    return elements
+
+
+def add_metadata_with_filetype(filetype: FileType):
+    def decorator(func: Callable):
+        @wraps(func)
+        def wrapper(*args, **kwargs):
+            elements = func(*args, **kwargs)
+            sig = inspect.signature(func)
+            params = dict(**dict(zip(sig.parameters, args)), **kwargs)
+            for param in sig.parameters.values():
+                if param.name not in params and param.default is not param.empty:
+                    params[param.name] = param.default
+            include_metadata = params.get("include_metadata", True)
+            if include_metadata:
+                metadata_kwargs = {
+                    kwarg: params.get(kwarg) for kwarg in ("include_page_breaks", "filename", "url")
+                }
+                return _add_element_metadata(
+                    elements,
+                    filetype=FILETYPE_TO_MIMETYPE[filetype],
+                    **metadata_kwargs,  # type: ignore
+                )
+            else:
+                return elements
+
+        return wrapper
+
+    return decorator